# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthTokenCreatedResponse": {
        "description": "Response that includes the full token (only on creation)",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "token",
          "created_at"
        ],
        "type": "object"
      },
      "AuthTokenResponse": {
        "description": "Response for listing (token is masked)",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "last_used_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "token_prefix": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "token_prefix",
          "created_at"
        ],
        "type": "object"
      },
...
      "CreateAuthToken": {
        "description": "DTO for creating a new token",
        "properties": {
          "description": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateProject": {
        "description": "DTO for creating a new project",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "JSON error response structure",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "OffsetPaginatedProjectResponse": {
        "description": "Paginated project response (offset-based) - concrete type for OpenAPI",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int64",
            "type": "integer"
          },
          "per_page": {
            "format": "int64",
            "type": "integer"
          },
          "total_count": {
            "format": "int64",
            "type": "integer"
          },
          "total_pages": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total_count",
          "page",
          "per_page",
          "total_pages"
        ],
        "type": "object"
      },
...
      "ProjectResponse": {
        "description": "Response with DSN included",
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "digested_event_count": {
            "format": "int32",
            "type": "integer"
          },
          "dsn": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "sentry_key": {
            "format": "uuid",
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "stored_event_count": {
            "format": "int32",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "sentry_key",
          "dsn",
          "stored_event_count",
          "digested_event_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "SortOrder": {
        "description": "Sort order direction",
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
...
      "UpdateProject": {
        "description": "DTO for updating a project",
        "properties": {
          "name": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "hex",
        "description": "40-character hex API token",
        "scheme": "bearer",
        "type": "http"
      },
      "sentry_auth": {
        "description": "Sentry SDK authentication header or ?sentry_key= query parameter",
        "in": "header",
        "name": "X-Sentry-Auth",
        "type": "apiKey"
      },
      "session_auth": {
        "in": "cookie",
        "name": "rustrak_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Rustrak Team"
    },
    "description": "Ultra-lightweight error tracking server compatible with Sentry SDKs",
    "license": {
      "name": "GPL-3.0"
    },
    "title": "Rustrak API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects",
        "parameters": [
          {
            "description": "Page number (1-indexed, default: 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Items per page (default: 20)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Sort order direction (default: desc = newest first)",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffsetPaginatedProjectResponse"
                }
              }
            },
            "description": "List of projects"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects - List projects with pagination",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Project created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict (name/slug taken)"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "POST /api/projects - Create a new project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "delete": {
        "operationId": "delete_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "DELETE /api/projects/{id} - Delete a project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Project details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{id} - Get a project by ID",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Project updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "PATCH /api/projects/{id} - Update a project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/tokens": {
      "get": {
        "operationId": "list_tokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuthTokenResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of tokens (masked)"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/tokens - List all tokens (masked)",
        "tags": [
          "tokens"
        ]
      },
      "post": {
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenCreatedResponse"
                }
              }
            },
            "description": "Token created (full token shown once)"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "POST /api/tokens - Create a new token",
        "tags": [
          "tokens"
        ]
      }
    },
    "/api/tokens/{id}": {
      "delete": {
        "operationId": "delete_token",
        "parameters": [
          {
            "description": "Token ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token revoked"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "DELETE /api/tokens/{id} - Revoke a token",
        "tags": [
          "tokens"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer_auth": []
    },
    {
      "session_auth": []
    },
    {
      "sentry_auth": []
    }
  ],
...
}